requirements.txt
openai>=0.27.0
python-dotenv>=0.19.0
azure-search-documents==11.4.0
azure-ai-textanalytics
argparse
pandas>=1.5.0
PyPDF2
streamlit>=1.24.0
openpyxl
azure-storage-blob
tiktoken
langchain>=0.0.200
langchain-community>=0.0.10
sqlalchemy>=1.4.0
psycopg2-binary
flask 
flask-cors
gunicorn